refactor: group bank notification parsing rules per source app
Banking apps each format their notifications differently, so parsing rules
are now grouped per source app instead of being assumed universal. The
PURCHASE / C2C pipe-format rules move into a dedicated Ameriabank parser
(com.banqr.ameriabank); a registry dispatches each notification to the
parser registered for its packageName, falling back to trying every known
parser when the source app is unknown (e.g. a manual paste).

Adding another bank later is just a new module under bankParsers/ — the
dispatcher and pipeline are untouched.

- bankParsers/ameriabank.js: the Ameria pipe parser + kind tables.
- bankParsers/index.js: registry + getParserForPackage().
- parseBankNotification.js: thin dispatcher; kindRequiresCategory now takes
  an optional packageName for per-app scoping.
- processBankNotifications / panel: pass packageName when checking
  kindRequiresCategory.
- Tests use the real com.banqr.ameriabank package and cover dispatch +
  unknown-source fallback. Docs updated.
