+++ app/services/notifications/parseBankNotification.js
@@ -1,274 +1,57 @@
 /**
- * Bank notification parser.
+ * Bank notification parser — source-app dispatcher.
  *
- * Banking apps post transaction notifications as a single pipe-delimited line,
- * e.g. the Ameria "ARCA transaction" notification:
+ * Parsing rules are grouped per banking app (see `bankParsers/`), because each
+ * bank formats its notifications differently. This module just routes a captured
+ * notification to the parser registered for its source app and exposes the
+ * per-app `kindRequiresCategory` helper.
  *
- *   title: АРКА транзакции
- *   text:  PURCHASE | 3,900.00 AMD | 4083***7027, | NAREK MEHRABYAN, AM | 28.06.2026 10:15 | BALANCE: 133,719.97 AMD
- *
- * This module turns that raw text into a structured, normalized object that the
- * rest of the app can map onto an operation. It is a *pure* function with no
- * side effects so it can be exhaustively unit-tested and reused regardless of
- * how notifications are ingested (polling, native events, manual paste).
- *
- * Design notes:
- * - Parsing is pattern-based, not strictly positional. Each pipe segment is
- *   classified by what it looks like (amount+currency, card mask, merchant,
- *   date/time, balance) so a reordered or slightly different layout still
- *   parses. The balance segment is deliberately ignored.
- * - Unrecognized text returns `null` so callers can cheaply skip the flood of
- *   non-transaction notifications the listener also sees.
+ * Today only Ameriabank (`com.banqr.ameriabank`) is supported. When a second
+ * bank is added it becomes another parser module in `bankParsers/` — nothing
+ * here changes.
  */
 
-/**
- * Maps a notification "kind" keyword to an operation type. Extend this table to
- * support more notification kinds (e.g. REFUND -> income) without touching the
- * parsing logic below.
- *
- * C2C is a client-to-client transfer (money sent to another person). It is an
- * expense like a purchase, but unlike a merchant purchase its category cannot be
- * inferred — see KINDS_REQUIRING_CATEGORY below.
- */
-const KIND_TO_TYPE = {
-  PURCHASE: 'expense',
-  C2C: 'expense',
-};
-
-const KNOWN_KINDS = Object.keys(KIND_TO_TYPE);
-
-/**
- * Kinds whose category must always be chosen by the user instead of being
- * inferred from learned merchant rules. A client-to-client transfer goes to the
- * same counterparty (a friend) for many different reasons, so any single learned
- * category would be wrong as often as right. These always land in the review
- * queue with the category left blank.
- */
-const KINDS_REQUIRING_CATEGORY = new Set(['C2C']);
-
-/**
- * Whether a notification kind must always have its category chosen manually.
- * @param {string} kind
- * @returns {boolean}
- */
-export const kindRequiresCategory = (kind) =>
-  KINDS_REQUIRING_CATEGORY.has(String(kind || '').toUpperCase());
-
-// "TO: N. DORVANYAN" — a leading recipient label on C2C transfer segments. The
-// descriptive part (the recipient) is what we keep as the merchant.
-const RECIPIENT_LABEL_RE = /^(?:TO|FROM)\s*:\s*/i;
-
-// "3,900.00 AMD" — amount with optional thousands separators followed by a
-// 3-letter currency code, anchored so "BALANCE: 133,719.97 AMD" never matches.
-const AMOUNT_CURRENCY_RE = /^([\d.,]+)\s+([A-Z]{3})$/;
-
-// "4083***7027" — a masked card number (4 digits, asterisks, 4 digits).
-const CARD_MASK_RE = /\d{2,6}\*{2,}\d{2,6}/;
-
-// "28.06.2026 10:15" — DD.MM.YYYY with an optional HH:MM time.
-const DATE_TIME_RE = /\b(\d{2})\.(\d{2})\.(\d{4})(?:\s+(\d{2}:\d{2}))?\b/;
-
-// A trailing ", AM" style ISO-3166 alpha-2 country code on the merchant segment.
-const TRAILING_COUNTRY_RE = /,\s*([A-Z]{2})\s*$/;
-
-/**
- * Normalize a localized amount string to a plain decimal string.
- *
- * Handles both grouping conventions without corrupting either:
- * - "1,234.56" (comma thousands, dot decimal) -> "1234.56"
- * - "1.234,56" (dot thousands, comma decimal) -> "1234.56"
- * - "12,50" (comma decimal)                   -> "12.50"
- *
- * When both separators are present, the one that appears last is the decimal
- * separator and the other is grouping. When only one separator is present it is
- * treated as a decimal point only if it is not a 3-digit group (so "1,234" and
- * "1.234" are read as thousands, while "12,50" is read as a decimal). The result
- * is a string so it feeds straight into the decimal currency layer without ever
- * becoming a lossy float.
- *
- * @param {string} raw - e.g. "3,900.00"
- * @returns {string|null} e.g. "3900.00", or null when no digits are present
- */
-const normalizeAmount = (raw) => {
-  if (!raw) return null;
-  let s = raw.replace(/[^\d.,]/g, '');
-  if (!/\d/.test(s)) return null;
-
-  const lastDot = s.lastIndexOf('.');
-  const lastComma = s.lastIndexOf(',');
-
-  if (lastDot !== -1 && lastComma !== -1) {
-    if (lastComma > lastDot) {
-      // Comma is the decimal separator: drop dot grouping, comma -> dot.
-      s = s.replace(/\./g, '').replace(',', '.');
-    } else {
-      // Dot is the decimal separator: drop comma grouping.
-      s = s.replace(/,/g, '');
-    }
-  } else if (lastComma !== -1) {
-    const parts = s.split(',');
-    if (parts.length === 2 && parts[1].length !== 3) {
-      s = s.replace(',', '.'); // single comma, not a 3-digit group -> decimal
-    } else {
-      s = s.replace(/,/g, ''); // grouping
-    }
-  } else if (lastDot !== -1) {
-    const parts = s.split('.');
-    if (parts.length > 2) {
-      s = s.replace(/\./g, ''); // multiple dots -> grouping
-    }
-    // single dot -> keep as the decimal point
-  }
-
-  return s;
-};
-
-/**
- * Convert "DD.MM.YYYY" to an ISO "YYYY-MM-DD" date string.
- *
- * Validates against a real calendar via a UTC round-trip, so impossible dates
- * like "31.02.2026" are rejected (returns null) rather than producing
- * "2026-02-31".
- *
- * @param {string} day
- * @param {string} month
- * @param {string} year
- * @returns {string|null} ISO date, or null if the date is not a real calendar date
- */
-const toIsoDate = (day, month, year) => {
-  const d = Number(day);
-  const m = Number(month);
-  const y = Number(year);
-  if (!Number.isInteger(d) || !Number.isInteger(m) || !Number.isInteger(y)) return null;
-  if (m < 1 || m > 12 || d < 1 || d > 31) return null;
-  const dt = new Date(Date.UTC(y, m - 1, d));
-  if (
-    dt.getUTCFullYear() !== y ||
-    dt.getUTCMonth() !== m - 1 ||
-    dt.getUTCDate() !== d
-  ) {
-    return null;
-  }
-  return `${year}-${month}-${day}`;
-};
+import { BANK_PARSERS, getParserForPackage } from './bankParsers';
 
 /**
  * Parse a captured bank notification into a normalized transaction descriptor.
  *
+ * Dispatch is by source app: the parser registered for `notification.packageName`
+ * handles it. When the package is unknown or missing (e.g. a manual paste), every
+ * registered parser is tried in turn and the first to recognize the format wins —
+ * each parser returns null for formats it doesn't handle, so this is safe.
+ *
  * @param {{ title?: string, text?: string, packageName?: string, postTime?: number }} notification
- * @returns {null | {
- *   kind: string,             // e.g. 'PURCHASE'
- *   type: 'expense'|'income', // operation type the kind maps to
- *   amount: string,           // normalized decimal string, e.g. '3900.00'
- *   currency: string,         // ISO code, e.g. 'AMD'
- *   cardMask: string|null,    // e.g. '4083***7027'
- *   merchant: string|null,    // e.g. 'NAREK MEHRABYAN'
- *   country: string|null,     // ISO alpha-2, e.g. 'AM'
- *   date: string|null,        // 'YYYY-MM-DD'
- *   time: string|null,        // 'HH:MM'
- *   requiresCategory: boolean,// true when the category must be chosen manually (C2C)
- *   packageName: string|null, // source app, passed through for rule scoping
- *   raw: string,              // the original text, kept for auditing/debugging
- * }}
+ * @returns {null | Object} a normalized descriptor (see a parser module for the
+ *   exact shape), or null when nothing recognizes the notification.
  */
 export const parseBankNotification = (notification) => {
   if (!notification || typeof notification.text !== 'string') return null;
 
-  const text = notification.text.trim();
-  if (!text) return null;
+  const parser = getParserForPackage(notification.packageName);
+  if (parser) return parser.parse(notification);
 
-  const segments = text
-    .split('|')
-    .map((segment) => segment.trim())
-    .filter(Boolean);
-  if (segments.length === 0) return null;
-
-  // 1. Kind — must be present for us to recognize this as a transaction.
-  const kindSegment = segments.find((segment) =>
-    KNOWN_KINDS.includes(segment.toUpperCase()),
-  );
-  if (!kindSegment) return null;
-  const kind = kindSegment.toUpperCase();
-  const type = KIND_TO_TYPE[kind];
-
-  // 2. Amount + currency — required; without it there's nothing to record.
-  let amount = null;
-  let currency = null;
-  for (const segment of segments) {
-    const match = segment.match(AMOUNT_CURRENCY_RE);
-    if (match) {
-      amount = normalizeAmount(match[1]);
-      currency = match[2];
-      break;
-    }
+  // Unknown/missing source app: fall back to trying every registered parser.
+  for (const candidate of BANK_PARSERS) {
+    const result = candidate.parse(notification);
+    if (result) return result;
   }
-  if (!amount || !currency) return null;
+  return null;
+};
 
-  // 3. Card mask — optional but the key to account binding.
-  const cardSegment = segments.find((segment) => CARD_MASK_RE.test(segment));
-  const cardMask = cardSegment
-    ? (cardSegment.match(CARD_MASK_RE) || [null])[0]
-    : null;
-
-  // 4. Date / time — pull from whichever segment carries a *valid* date. Keep
-  //    scanning if a segment's date is impossible (e.g. 31.02.2026) so a valid
-  //    date elsewhere is still found.
-  let date = null;
-  let time = null;
-  for (const segment of segments) {
-    const match = segment.match(DATE_TIME_RE);
-    if (match) {
-      const iso = toIsoDate(match[1], match[2], match[3]);
-      if (iso) {
-        date = iso;
-        time = match[4] || null;
-        break;
-      }
-    }
-  }
-
-  // 5. Merchant + country — the descriptive segment that is none of the above.
-  //    It carries letters, isn't the kind, balance, card, amount or date.
-  let merchant = null;
-  let country = null;
-  for (const segment of segments) {
-    if (segment === kindSegment || segment === cardSegment) continue;
-    if (/^BALANCE\b/i.test(segment)) continue;
-    if (AMOUNT_CURRENCY_RE.test(segment)) continue;
-    if (DATE_TIME_RE.test(segment)) continue;
-    if (!/[A-Za-zÀ-ɏЀ-ӿ]/.test(segment)) continue;
-
-    let candidate = segment;
-    const countryMatch = candidate.match(TRAILING_COUNTRY_RE);
-    if (countryMatch) {
-      country = countryMatch[1];
-      candidate = candidate.replace(TRAILING_COUNTRY_RE, '').trim();
-    }
-    candidate = candidate.replace(/,\s*$/, '').trim();
-    // Strip a "TO:"/"FROM:" recipient label so a C2C transfer descriptor keeps
-    // just the counterparty's name as its merchant/description.
-    candidate = candidate.replace(RECIPIENT_LABEL_RE, '').trim();
-    if (candidate) {
-      merchant = candidate;
-      break;
-    }
-  }
-
-  return {
-    kind,
-    type,
-    amount,
-    currency,
-    cardMask,
-    merchant,
-    country,
-    date,
-    time,
-    // C2C transfers must always be reviewed so the user picks the category.
-    requiresCategory: kindRequiresCategory(kind),
-    packageName: notification.packageName || null,
-    raw: text,
-  };
+/**
+ * Whether a notification kind must always have its category chosen manually
+ * (e.g. C2C transfers). Scoped to the source app when known; otherwise true if
+ * any registered parser treats the kind that way.
+ *
+ * @param {string} kind
+ * @param {string} [packageName] - source app, for per-app scoping
+ * @returns {boolean}
+ */
+export const kindRequiresCategory = (kind, packageName) => {
+  const parser = getParserForPackage(packageName);
+  if (parser) return parser.kindRequiresCategory(kind);
+  return BANK_PARSERS.some((candidate) => candidate.kindRequiresCategory(kind));
 };
 
 export default parseBankNotification;
